Filtering msfragger Data

Filtering msfragger data. In this implementation, the peptide-level filter optimizes both ppm and one of Expectation or PeptideProphet Probability thresholds to achieve maximum number of peptide identifications within a given FDR constraint. The accession-level filter optimizes based on peptides_per_1000aa, so compute_num_peptides_per_1000aa must be used first.

Usage

filter_msfragger_data(
  msnid,
  level,
  filtering_criterion = c("pp_prob", "evalue"),
  fdr.max = 0.01,
  n.iter.grid = 500,
  n.iter.nm = 100
)

Arguments

msnid

(MSnID object) collated msfragger output

level

(character) Level at which to perform FDR filter. The name of a column in psms(msnid). Currently, only "peptide" or "accession" are supported. The added level SiteID makes sense only for PTM data and first requires mapping of the modification site using MSnID::map_mod_sites.

filtering_criterion

(character) One of "evalue" which is expectation value or "pp_prob" - peptide prophet probability. Default is "pp_prob".

fdr.max

(numeric) Maximum acceptable FDR. Default is 0.01 (1%).

n.iter.grid

(numeric) number of grid-distributed evaluation points.

n.iter.nm

(numeric) number of iterations for Nelder-Mead optimization algorithm.

...

arguments passed to filter_msfragger_data.

Value

(MSnID object) filtered msfragger output

See also