Wrapper for PlexedPiper

A wrapper function for PlexedPiper that performs all the steps necessary to go from MS-GF+ and MASIC data to Reporter Ion Intensity (RII) and Ratio Results tables used by the MoTrPAC Bioinformatics Center (BIC).

Usage

run_plexedpiper(
  msgf_output_folder,
  fasta_file,
  masic_output_folder,
  ascore_output_folder = NULL,
  proteomics,
  study_design_folder,
  species,
  annotation,
  file_prefix = NULL,
  unique_only = FALSE,
  global_results = NULL,
  refine_prior = FALSE,
  write_results_to_file = TRUE,
  output_folder = NULL,
  save_env = FALSE,
  return_results = FALSE,
  verbose = TRUE
)

Arguments

msgf_output_folder

(character) Path to MSGF+ results folder

fasta_file

(character) Path to FASTA file

masic_output_folder

(character) MASIC results folder

ascore_output_folder

(character) AScore results folder

proteomics

(character) Either "pr" - proteomics, "ph" - phosphoproteomics, "ub" - ubiquitinomics, or "ac" - acetylomics

study_design_folder

(character) Folder containing the three study design tables: fractions.txt, samples.txt, and references.txt

species

(character) Scientific name of species (e.g. "Rattus norvegicus", "Homo sapiens", etc.)

annotation

(character) Source for annotations: either RefSeq or UniProt (case insensitive).

file_prefix

(character) Prefix for the file name outputs.

unique_only

(logical) Whether to discard peptides that match multiple proteins in the parsimonious protein inference step. Default FALSE.

global_results

(character) Only for PTM experiments. Ratio results from a global protein abundance experiment. If provided, prioritized inference will be performed. Otherwise, parsimonious inference is performed without a prior. See infer_parsimonious_accessions for details.

refine_prior

(logical) if FALSE (default), peptides are allowed to match multiple proteins in the prior. That is, the greedy set cover algorithm is only applied to the set of proteins not in the prior. If TRUE, the algorithm is applied to the prior and non-prior sets separately before combining. See infer_parsimonious_accessions for more details.

write_results_to_file

(logical) Whether to write the results to files.

output_folder

(character) Output folder name to save results. If not provided, it will save it to the current directory.

save_env

(logical) Whether to save the R environment to the output folder.

return_results

(logical) Whether to return the ratio and rii results.

verbose

(logical) Whether to show messages describing steps.

Value

(list) If return_results is TRUE, it returns list with ratio and RII data frames.

Examples

if (FALSE) {
results <- run_plexedpiper(msgf_output_folder = "~/path/to/msgfplus/",
                           fasta_file  = "~/path/to/fasta/sequence.fasta",
                           masic_output_folder = "~/path/to/masic-results/",
                           ascore_output_folder = "~/path/to/ascore-results/",
                           proteomics = "ph",
                           study_design_folder = "~/path/to/study-design/",
                           species = "Rattus norvegicus",
                           annotation = "RefSeq",
                           global_results = "~/path/to/global/ratio.txt",
                           unique_only = FALSE,
                           refine_prior = FALSE,
                           output_folder = "~/path/to/pp-results/",
                           file_prefix = "msgfplus-pp-results",
                           return_results = TRUE,
                           verbose = TRUE)
}